+++ tests/error/05_null_test.c
@@ -2,9 +2,9 @@
 
 int	null_test05(void)
 {
-	t_unit_test *testlist;
-	int backup;
-	int pid;
+	t_unit_test	*testlist;
+	int			backup;
+	int			pid;
 
 	testlist = NULL;
 	backup = error_test_constructor();
