+++ tests/error/08_null_test.c
@@ -2,9 +2,9 @@
 
 int	null_test08(void)
 {
-	t_unit_test *testlist;
-	int backup;
-	int pid;
+	t_unit_test	*testlist;
+	int			backup;
+	int			pid;
 
 	testlist = NULL;
 	backup = error_test_constructor();
