+++ tests/error/basic_tests_error.h
@@ -31,7 +31,6 @@ int	error_launcher(void);
 int	error_test_constructor(void);
 int	error_test_destructor(int backup);
 int	error_test(void);
-int	null_test00(void);
 int	null_test01(void);
 int	null_test02(void);
 int	null_test03(void);
@@ -42,5 +41,6 @@ int	null_test07(void);
 int	null_test08(void);
 int	null_test09(void);
 int	null_test10(void);
+int	null_test11(void);
 
 #endif
